dat science/requirements.txt
# Data Collection
tweepy==4.14.0
selenium==4.11.2
beautifulsoup4==4.12.2
requests==2.31.0

# Data Processing
pandas==2.1.0
numpy==1.24.3
python-dotenv==1.0.0

# NLP & ML
transformers==4.33.2
torch==2.0.1
scikit-learn==1.3.0
spacy==3.6.1
nltk==3.8.1

# Visualization & Dashboard
plotly==5.16.1
dash==2.13.0
dash-bootstrap-components==1.5.0

# API & Backend
fastapi==0.103.1
uvicorn==0.23.2
pymongo==4.5.0

# Testing & Development
pytest==7.4.2
black==23.7.0
flake8==6.1.0
jupyter==1.0.0
notebook==7.0.3 